% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn_ramp.R
\name{asn_cols}
\alias{asn_cols}
\title{Return ASN Brand Identity hex colors by position}
\usage{
asn_cols(...)
}
\arguments{
\item{...}{Returns all ASN Brand Colors or use an integer 1-12}
}
\value{
A hex color code
}
\description{
Return ASN Brand Identity hex colors by position
}
\examples{
asn_cols()
asn_cols(1)
}
