% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{asn_dark_mode}
\alias{asn_dark_mode}
\title{Light/Neon Colors for Black Background}
\usage{
asn_dark_mode()
}
\value{
Light/Neon Colors for Black Background
}
\description{
Light/Neon Colors for Black Background
}
