% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{asn_main_blue}
\alias{asn_main_blue}
\title{White and Main Blue for sequential palette}
\usage{
asn_main_blue()
}
\value{
White and Main Blue for sequential palette
}
\description{
White and Main Blue for sequential palette
}
