% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{asn_main_orange}
\alias{asn_main_orange}
\title{White and Orange for sequential palette}
\usage{
asn_main_orange()
}
\value{
White and Orange for sequential palette
}
\description{
White and Orange for sequential palette
}
