% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn_ramp.R
\name{asn_pal}
\alias{asn_pal}
\title{Create Palettes for Use in ggplot2 data viz with interpolation}
\usage{
asn_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of 9 color Schemes using ASN Brand Identity colors}

\item{reverse}{Should the color orders of the palette be reversed?}

\item{...}{Additional arguments to colorRampPalette}
}
\value{
A function to interpolate between the two outer values in the color scheme
}
\description{
Create Palettes for Use in ggplot2 data viz with interpolation
}
\examples{
asn_pal()
}
