% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{asn_secondary_blue}
\alias{asn_secondary_blue}
\title{White and Secondary Blue for sequential palette}
\usage{
asn_secondary_blue()
}
\value{
White and Secondary Blue for sequential palette
}
\description{
White and Secondary Blue for sequential palette
}
