% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_section_label.R
\name{asn_section_label}
\alias{asn_section_label}
\title{Create a consistent-width section label}
\usage{
asn_section_label(label = "00 Libraries", typst = FALSE)
}
\arguments{
\item{label}{The section name, typically prefixed with section number.}

\item{typst}{Logical. Changes the comment character for section labels in a Typst document. Default FALSE.}
}
\value{
A character string of the section label, invisibly. Also copied to the clipboard in interactive sessions.
}
\description{
Creates a custom commented section label for outlining sections in your script that is copied to the clipboard for pasting in R code and Typst/Quarto documents.
}
\examples{
\dontrun{
  asn_section_label("01 Data")
}
}
