% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_theme.R
\name{asn_theme}
\alias{asn_theme}
\title{ASN Brand Identity for \code{highcharter} Package}
\usage{
asn_theme
}
\value{
Custom highcharter theme with ASN Brand Identity
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{asn_theme} has been superseded in favor of \code{hc_theme_asn}, which follows the
\code{highcharter} naming convention.
}
\examples{
library(highcharter)
highchart() |>
  hc_add_series(
    data = iris,
    type = "scatter",
    hcaes(
      x = Sepal.Length,
      y = Sepal.Width,
      group = Species
    )
  ) |>
  hc_add_theme(
    asn_theme
  )
}
