% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{board_prep_df}
\alias{board_prep_df}
\title{Board Prep Resources}
\format{
A data frame with 8 columns
\enumerate{
\item \code{q6}: Repondent medical school location, either "United States" or "Other country".
\item \code{q65_1}: ASN Kidney Self-Assessment Program (KSAP) selected as a used resource (\if{html}{\out{<NA>}} = "No").
\item \code{q65_2}: ASN Nephrology Self-Assessment Program (NephSAP) selected as a used resource (\if{html}{\out{<NA>}} = "No").
\item \code{q65_3}: ASN Board Review Course & Update selected as a used resource (\if{html}{\out{<NA>}} = "No").
\item \code{q65_4}: Oakstone/Brigham Intensive Review of Nephrology Course (\if{html}{\out{<NA>}} = "No").
\item \code{q65_5}: Other board review course (\if{html}{\out{<NA>}} = "No").
\item \code{q65_6}: Other resource(s) (please specify) (\if{html}{\out{<NA>}} = "No").
\item \code{q65_6_text}: Free text response describing other resource(s) used.
}
}
\source{
\url{https://data.asn-online.org/posts/2025_fellow_survey/}
}
\usage{
board_prep_df
}
\description{
Resources used by nephrology fellows in preparation for initial board certication in
nephrology. This is a synthesized data set generated based on proportions of actual responses to
measures in the 2025 ASN Nephrology Fellow Survey.
}
\examples{
board_prep_df

}
\keyword{datasets}
