% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R, R/get_labs.R
\name{get_labels}
\alias{get_labels}
\title{Retrieve labels from labeled data frames derived from SPSS \code{.sav} files}
\usage{
get_labels(x)

get_labels(x)
}
\arguments{
\item{x}{A labelled R object}
}
\value{
Labels for the R object

Labels for the R object
}
\description{
Convenience function to extract labels from labelled data frames or objects,
usually from survey data captured in Qualtrics imported into R using the \code{haven} package.
}
\examples{
label_example <-
  structure(
    data.frame(
      Q1 = c("U.S.", "Canada", "U.S.", "Germany", "U.S.")
    ),
    label = "What country do you live in?"
  )

get_labels(label_example)
label_example <-
  structure(
    data.frame(
      Q1 = c("U.S.", "Canada", "U.S.", "Germany", "U.S.")
    ),
    label = "What country do you live in?"
  )

get_labels(label_example)
}
