% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_df}
\alias{grid_df}
\title{Grid Data Frame for asn_tilegrid()}
\format{
A data frame with 6 columns:
\enumerate{
\item \code{State}: USPS codes for U.S. states and the District of Columbia.
\item \code{x}: Corresponding x coordinate for tilegrid.
\item \code{y}: Corresponding y coordinate for tilegrid.
}
}
\source{
\url{https://github.com/kpivert/economist_choropleth}
}
\usage{
grid_df
}
\description{
Data frame of The Economist's 8 x 11 rectangular tile grid for the asn_tilegrid() function, adapted from the tilegrid
choropleth designed by The Economist Data Team.
}
\examples{
grid_df

}
\keyword{datasets}
