% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_theme_asn.R
\name{hc_theme_asn}
\alias{hc_theme_asn}
\title{ASN Brand Identity for \code{highcharter} Package}
\usage{
hc_theme_asn
}
\value{
Custom \code{highcharter} theme with ASN Brand Identity
}
\description{
ASN Brand Identity for \code{highcharter} Package
}
\examples{
library(highcharter)

highchart() |>
  hc_add_series(
    data = iris,
    type = "scatter",
    hcaes(
      x = Sepal.Length,
      y = Sepal.Width,
      group = Species
    )
  ) |>
  hc_add_theme(
    hc_theme_asn
  )
}
