% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neph_per_100K_df}
\alias{neph_per_100K_df}
\title{Nephrologists per 100K U.S. Population by State}
\format{
A data frame with 2 columns:
\enumerate{
\item \code{MailState}: USPS codes for U.S. states and the District of Columbia.
\item \code{neph_per_100k}: Ratio of practicing nephrologists per 100K adult population in state/DC.
}
}
\source{
\url{https://www.census.gov/data/datasets/time-series/demo/popest/2020s-counties-detail.html}
}
\usage{
neph_per_100K_df
}
\description{
Ratio of practicing nephrologists per 100K adult population per state and District of
Columbia in 2023. Nephrologist counts aggregated from the American Medical Assosciation
Physician Professional Database. U.S. adult population extracted from the U.S. Census Bureau.
}
\examples{
neph_per_100K_df

}
\keyword{datasets}
