% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_functions.R
\name{pct_chg}
\alias{pct_chg}
\title{Function to calculate percentage change from previous observation}
\usage{
pct_chg(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Function to calculate percentage change from previous observation
}
\examples{
library(dplyr)
x <- 1:5
pct_chg(x)
}
