% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_functions.R
\name{pct_fun}
\alias{pct_fun}
\title{Function to create ordered frequency table with percentages for each level}
\usage{
pct_fun(data, col_var, likert = FALSE, accuracy = 1)
}
\arguments{
\item{data}{A data frame}

\item{col_var}{Column of interest to calculate ordered frequency table}

\item{likert}{Is variable a Likert scale question/item? Will retain scale order if TRUE.}

\item{accuracy}{Round to how many decimal places for Percent column. Default = 1.}
}
\value{
A data frame
}
\description{
Function to create ordered frequency table with percentages for each level
}
\examples{
pct_fun(iris, col_var = Species)
}
