% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_functions.R
\name{pct_norm}
\alias{pct_norm}
\title{Function to calculate percentage change from first observation in vector}
\usage{
pct_norm(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Function to calculate percentage change from first observation in vector
}
\examples{
library(dplyr)
x <- 1:5
pct_norm(x)
}
