% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_label.R
\name{pretty_label}
\alias{pretty_label}
\title{Create a consistent-width section label}
\usage{
pretty_label(label = "00 Libraries")
}
\arguments{
\item{label}{The section name, typically prefixed with section number.}
}
\value{
A line of unquoted text that can be copied from the console and pasted in the source file.
}
\description{
Create a consistent-width section label
}
\examples{
pretty_label("01 Data")
}
