% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{scale_color_asn}
\alias{scale_color_asn}
\title{Create scale_color_ Palettes for Use in ggplot2 data viz (no interpolation)}
\usage{
scale_color_asn(..., palette = "primary")
}
\arguments{
\item{...}{Additional arguments to ggplot2::scale_color_manual()}

\item{palette}{One of 9 color Schemes using ASN Brand Identity colors}
}
\value{
scale_color_asn() function
}
\description{
Create scale_color_ Palettes for Use in ggplot2 data viz (no interpolation)
}
\examples{
library(ggplot2)
ggplot(
  iris,
  aes(
    x = Sepal.Length,
    y = Petal.Length,
    color = Species
   )
 ) +
 geom_point() +
 geom_smooth(
   method = "lm",
   se = FALSE
 ) +
 scale_color_asn(palette = "dark_mode") +
 theme(
   panel.background = element_rect(color = "#000000", fill = "#000000"),
   plot.background = element_rect(color = "#000000", fill = "#000000"),
   panel.grid = element_blank()
   )
}
