% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn_ramp.R
\name{scale_color_asn_ramp}
\alias{scale_color_asn_ramp}
\title{Use ASN Brand Identity interpolated color scale}
\usage{
scale_color_asn_ramp(
  palette = "primary",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{One of 9 color Schemes using ASN Brand Identity colors ("primary, secondary, bw, main_blue, secondary_blue, main_orange, highlight_one_of_two, dark_mode, spring")}

\item{discrete}{Is the variable discrete?}

\item{reverse}{Should the color orders of the palette be reversed?}

\item{...}{Additional arguments to ggplot2::discrete_scale() constructor}
}
\value{
scale_color_asn_ramp() function
}
\description{
Use ASN Brand Identity interpolated color scale
}
\examples{
library(ggplot2)
ggplot(
  iris,
  aes(
    x = Sepal.Length,
    y = Petal.Length,
    color = Species
   )
 ) +
 geom_point() +
 geom_smooth(
   method = "lm",
   se = FALSE
 ) +
 scale_color_asn_ramp()
}
