% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R
\name{scale_fill_asn}
\alias{scale_fill_asn}
\title{Create scale_fill_ Palettes for Use in ggplot2 data viz (no interpolation)}
\usage{
scale_fill_asn(..., palette = "primary")
}
\arguments{
\item{...}{Additional arguments to ggplot2::scale_fill_manual()}

\item{palette}{One of 9 color Schemes using ASN Brand Identity colors}
}
\value{
scale_fill_asn() function
}
\description{
Create scale_fill_ Palettes for Use in ggplot2 data viz (no interpolation)
}
\examples{
library(ggplot2)
ggplot(
  chickwts,
  aes(
    x = feed,
    y = weight,
    fill = feed
   )
 ) +
 geom_boxplot() +
 scale_fill_asn(palette = "secondary")
}
