% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn_ramp.R
\name{scale_fill_asn_ramp}
\alias{scale_fill_asn_ramp}
\title{Use ASN Brand Identity interpolated fill scale}
\usage{
scale_fill_asn_ramp(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of 9 color Schemes using ASN Brand Identity colors ("primary, secondary, bw, main_blue, secondary_blue, main_orange, highlight_one_of_two, dark_mode, spring")}

\item{discrete}{Is the variable discrete?}

\item{reverse}{Should the color orders of the palette be reversed?}

\item{...}{Additional arguments to ggplot2::discrete_scale() constructor}
}
\value{
scale_fill_asn_ramp() function
}
\description{
Use ASN Brand Identity interpolated fill scale
}
\examples{
library(ggplot2)
ggplot(
  chickwts,
  aes(
    x = feed,
    y = weight,
    fill = feed
   )
 ) +
 geom_boxplot() +
 scale_fill_asn_ramp(palette = "main_orange")
}
