% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_all.R
\name{select_all}
\alias{select_all}
\title{Count the number of unique responses to "Select All" Questions}
\usage{
select_all(data, group_var = NULL, column_prefix)
}
\arguments{
\item{data}{A data frame, typically from a Qualtrics Survey, where responses to 'select all' questions are provided
in consecutive columns prefixed with the question number and underscore, with a number
corresponding to each response option.}

\item{group_var}{Column name for aggregrating variable of interest if the
results should the results be grouped by another variable.}

\item{column_prefix}{Question number and underscore}
}
\value{
A dataframe.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{select_all()} has been superseded in favor of \code{check_all()} because of a name
space conflict with the deprecated \code{dplyr::select_all()}. For backward compatability
prefix the function with the package name: \code{asn::select_all()}.
}
\examples{
test_df <-
  dplyr::tribble(
    ~Q4, ~Q15_1, ~Q15_2, ~Q15_3, ~Q15_4,
    "Community", NA, NA, NA, NA,
    "Academic", NA, "TREKS", NA, NA,
    "Academic", NA, NA, "Lipps Fellowship", NA,
    "Community", "STARS", NA, NA, "Campbell Fellows",
    "Academic", "STARS", NA, NA, "Campbell Fellows",
    "Community", NA, "TREKS", NA, NA
   ) |>
  dplyr::mutate(
    dplyr::across(dplyr::everything(), as.factor)
)
select_all(
  data = test_df,
  group_var = Q4,
  column_prefix = "Q15"
)
}
