% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_ggplot2_themes.R
\name{theme_asn_dark}
\alias{theme_asn_dark}
\title{Custom \code{ggplot2} dark theme compliant with ASN's brand identity.}
\usage{
theme_asn_dark(base_size = 15, title_font = "Gotham", base_font = "Roboto")
}
\arguments{
\item{base_size}{Base font size in points (pts)}

\item{title_font}{Typeface for plot titles, subtitles, and faceted titles.
Defaults to \code{Gotham} which is the primary typeface for ASN's brand identity.}

\item{base_font}{Typeface for other plot text, including axes titles and captions.
Defaults to \code{Roboto}.}
}
\value{
A dark-mode version of ASN brand identity-compliant \code{ggplot2} theme, built off of \code{theme_minimal()}.
}
\description{
Custom \code{ggplot2} dark theme compliant with ASN's brand identity.
}
\examples{
library(ggplot2)

ggplot(
  iris,
  aes(
    x = Sepal.Length,
    y = Sepal.Width,
    color = Species
    )
 ) +
 geom_point() +
 geom_smooth(
   method = "lm"
 ) +
 labs(
   x = "Length",
   y = "Width",
   title = "Relationship Between Iris Length and Width",
   subtitle = "<i>Setosa</i>, <i>Versicolor</i>, & <i>Virginica</i> are positive"
   ) +
 facet_wrap(~ Species) +
 theme_asn_dark(
   title_font = "serif",
   base_font = "sans"
 ) +
 scale_color_asn(palette = "dark_mode")
}
