% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn-package.R
\docType{package}
\name{asn-package}
\alias{asn}
\alias{asn-package}
\title{asn: ASN Brand Identity and Helper Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Custom ggplot2 and highcharter functions for implementing ASN's brand identity in data visualizations. Additional functions for calculating summaries for 'select all' survey questions and calculating percent changes are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ASNDataAnalytics/asn}
  \item \url{https://asndataanalytics.github.io/asn/}
  \item Report bugs at \url{https://github.com/ASNDataAnalytics/asn/issues}
}

}
\author{
\strong{Maintainer}: Kurtis Pivert \email{kpivert@asn-online.org} (\href{https://orcid.org/0000-0002-3428-2427}{ORCID})

}
\keyword{internal}
