% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn.R, R/scale_asn_ramp.R
\docType{data}
\name{asn_colors}
\alias{asn_colors}
\title{ASN Brand Identity for ggplot2 (no interpolation)}
\format{
An object of class \code{character} of length 12.

An object of class \code{character} of length 12.
}
\usage{
asn_colors

asn_colors
}
\description{
ASN brand colors and palettes for use with ggplot2.

ASN brand colors and palettes for use with ggplot2.
}
\keyword{datasets}
