% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_asn_ramp.R
\docType{data}
\name{asn_colors}
\alias{asn_colors}
\title{ASN Brand Identity with Interpolation for ggplot2}
\format{
An object of class \code{character} of length 12.
}
\usage{
asn_colors
}
\description{
ASN brand colors and palettes for use with ggplot2.
}
\keyword{datasets}
