% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_theme.R
\docType{data}
\name{asn_theme}
\alias{asn_theme}
\title{ASN Brand Identity for highcharter Package}
\format{
An object of class \code{hc_theme} of length 8.
}
\usage{
asn_theme
}
\value{
Custom highcharter theme with ASN Brand Identity
}
\description{
ASN Brand Identity for highcharter Package
}
\examples{
library(highcharter)
highchart() |>
  hc_add_series(
    data = iris,
    type = "scatter",
    hcaes(
      x = Sepal.Length,
      y = Sepal.Width,
      group = Species
    )
  ) |>
  hc_add_theme(
    asn_theme
  )
}
\keyword{datasets}
