% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_tilegrid.R
\name{asn_tilegrid}
\alias{asn_tilegrid}
\title{Create a U.S. Tilegrid Choropleth Map in ggplot2 in style of The Economist}
\usage{
asn_tilegrid(
  data = df,
  data_source = "Source",
  plot_title = "Plot Title",
  state_column = "state_column_name",
  numeric_column = n,
  plot_panel_color = "#ffffff",
  palette = c("#FCC8B4", "#FAAA90", "#F58469", "#EF1B26"),
  breaks = c(1, 2, 6, 11),
  zero_color = NULL,
  na_color = NULL
)
}
\arguments{
\item{data}{A data frame, comprising a column with 2-letter USPS state abbreviations and a column with a
numeric variable to be visualized in a tilegrid choropleth map.}

\item{data_source}{Source of the numeric data visualized in the choropleth.}

\item{plot_title}{Plot title.}

\item{state_column}{Column containing 2-letter USPS state abbreviations.}

\item{numeric_column}{Column containing numeric data being visualized in choropleth.}

\item{plot_panel_color}{Background color for plot.}

\item{palette}{Vector of hex colors mapping numeric column to choropleth map fill color.}

\item{breaks}{Numeric vector of breaks corresponding to colors in the color palette mapping to the numeric column.}

\item{zero_color}{Optional hex color to indicate states with a 0 numeric value.}

\item{na_color}{Optional hex color to indicate states with missing values.}
}
\value{
ggplot2 object
}
\description{
Create a U.S. Tilegrid Choropleth Map in ggplot2 in style of The Economist
}
\examples{
asn_tilegrid(
  data = neph_per_100K_df,
  data_source = "AMA PPD and U.S. Census Bureau.",
  plot_title = "Ratio of Nephrologists per 100K Adult Population",
  state_column = "MailState",
  numeric_column = neph_per_100k,
  plot_panel_color = "#ffffff",
  palette = c("#f0f8ff", "#a0bdd8", "#5081b2", "#00468b"), 
  breaks = c(2.21, 3.1, 4, 5.6), 
  zero_color = NULL,
  na_color = "#EEECE6"
)
}
