% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labs.R
\name{get_labs}
\alias{get_labs}
\title{Retrieve labels from labeled data frame from SPSS .sav Files}
\usage{
get_labs(x)
}
\arguments{
\item{x}{A labelled R object}
}
\value{
Labels from R object
}
\description{
Uses purrr's attr_getter() function to create a function to extract labels,m
usually questions from Qualtrics data loaded into R from haven.
}
\examples{
label_example <- structure(
  c(1:4),
  label = "I'm a label"
  )
get_labs(label_example)
}
