% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labs.R
\name{get_labs}
\alias{get_labs}
\title{Retrieve labels from labeled data frames from SPSS \code{.sav} files}
\usage{
get_labs(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{get_labs()} was renamed to \code{get_labels()} to avoid a namespace conflict with the \code{ggplot2::get_labs()}
function introduced in version 3.5.2.
}
\keyword{internal}
