% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_functions.R
\name{pct_fun}
\alias{pct_fun}
\title{Function to create ordered frequency table with percentages for each level}
\usage{
pct_fun(data = dat, col_var)
}
\arguments{
\item{data}{A data frame}

\item{col_var}{Column of interest to calculate ordered frequency table}
}
\value{
A data frame
}
\description{
Function to create ordered frequency table with percentages for each level
}
\examples{
pct_fun(iris, col_var = Species)
}
